Implement symlink system to handle different local IPs
By modifying wp-config.php to be different depending on the system, each user can create a symbolic link for wp-config.php
(which doesn't exist), and wordpress will use the IP you specify. This is needed because wordpress needs to know the host
in order to redirect to other pages, but this host is different depending on if you use Windows, Linux (as GCP will), or
Mac. Hopefully now the database can be imported without affecting the IP.